+++ src/plonemeeting/portal/core/utils.py
@@ -22,17 +22,20 @@ def get_api_url_for_meetings(institution, meeting_UID=None):
     )
     if meeting_UID:
         url = "{0}&UID={1}&fullobjects=True".format(url, meeting_UID)
+    else:
+        url = "{0}{1}".format(url, institution.additional_meeting_query_string_for_list)
     return url
 
 
 def get_api_url_for_meeting_items(institution, meeting_UID):
     if not institution.plonemeeting_url or not institution.meeting_config_id:
         return
-    url = "{0}/{1}?getConfigId={2}&linkedMeetingUID={3}&fullobjects=True".format(
+    url = "{0}/{1}?getConfigId={2}&linkedMeetingUID={3}&fullobjects=True{4}".format(
         institution.plonemeeting_url.rstrip("/"),
         PLONEMEETING_API_MEETING_ITEMS_VIEW,
         institution.meeting_config_id,
         meeting_UID,
+        institution.additional_published_items_query_string,
     )
     return url